Move validation classes to own files. Rename too.
